Lift `(Compile, compile)` restriction of SBT's `scalacOptions` in README.md
Now, that we have #85, maybe it might be worth considering to drop the
`in (Compile, compile)` from SBT settings?

However, some *new* projects (the ones that'll use the README.md for
reference) might want to use warts in, say, tests... But perhaps it
would be wiser for them not to? (=