Fixes #215 Optionally exclude macro expansions from checks
Introduces a `no-macros` plugin argument and `wartremoverNoMacros`
setting which is consulted at the same point as suppression
annotations.

That looks like the cleanest way to short-circuit traversal.

The config flag is communicated through a member of `WartUniverse`,
but the traverser could also look up the plugin through `global`
if it's desirable to keep the `WartUniverse` from expanding.

A mechanism is required for automated testing with different
plugin configurations. I tested manually on a trivial project
that uses the f-interpolator, under 2.11.8 and 2.10.5.